Prevent usage on drop_caches=3 after changes to FS for new shrink API.
There is a problem with drop_caches=3, in some unknown conditions,

sysctl -w vm.drop_caches=3 will be stuck and cant be unlocked.
CPU will be max on all cores till reboot.

possible bug in F2FS, as i dont see that problem with my device using EXT4

or it's something else. there is no log about any problem in dmesg.
So... for now just block that option.

MM has to be updated much more to resolve this. but then i loose
Adaptive LMK and VM pressure code made for 3.10.y
I dont like that.
it's will be best to stay with 3.10.y structure with few patches from main line.